backend/requirements.txt
fastapi
uvicorn[standard]
pydantic
python-dotenv
google-generativeai
firebase-admin
pyrebase4
python-jose[cryptography]
passlib[bcrypt]
google-cloud-firestore